pyproject.toml
[tool.poetry]
name = "otter-welcome-buddy"
version = "0.1.0"
description = ""
authors = ["proyecto-nutria <contact@proyectonutria.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"
APScheduler = "^3.9.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
flake8 = "^4.0.1"
pre-commit = "^2.20.0"
pytest-asyncio = "^0.19.0"
pytest-mock = "^3.10.0"

[tool.pylint."MESSAGES CONTROL"]
disable = ["E0401", "C0114"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
